nano-gui - Add command and read function
This allows for reading a byte after sending a command since a few commands can return data.
Since this eink display is a 3 wire SPI this will only work if we use a soft spi implementation to manually bang out clock cycles w/o putting data on the MOSI pin or if a resistor is used and this is handled in hardware.
Therefore, this update adds optional support for the software solution that the client can use if desired.