nano-gui - Add work-around to non-updating display issue
Sometimes attempting to update the display with the show() function results in nothing being updated on the display. When that happens, I found that a change in the display's status flag. When this issue happens, one or more of these status bits always is set. So, adding a check for it here and a re-initialization of the display to mitigate. Just re-initializing the display did not seem to resolve this issue when it occurs, but putting the display to sleep and then powering it back up and initializing it, aka turning it off and on again, does resolve this issue.