nano-gui - Remove work-around and fix CS usage
I think the issue I was having that I tried to solve with the status work-around was that I wasn't setting the chip select pin at the end of the refresh data transfer. After adding this, testing so far hasn't had any issues. So for now, I'm removing the complexity of the status work-around and sticking with this.

NOTE: When using the soft-spi work-around, globally configuring the soft-spi pins screws up their usage on the regular hardware spi driver. So, I would have had to work-around that as well.