nano-gui - Use a tx buffer instead of sending data one byte at a time
When refreshing the screen, the existing method is to send the frame buffer contents out one byte at a time. This adds significant overhead; the runtime of the show() function is around ~260msec.
Instead, building up the frame buffer contents into a dedicated transmit buffer and then sending that whole tx buffer in one SPI transaction dramatically increases performance at the cost of RAM usage. Using a TX buffer drops the runtime of the show() function to about ~40msec.