protocols: SerialProtocol - Add ability to parse non-aligned packets
Previous implementation assumed that the protocol would receive individual SerialPackets on every receive call. Since the underlaying transport doesn't guarantee that and, more importantly, since this is a serial protocol we need to handle any random blob of serial bytes received and correctly parse out SerialPackets. This update adds that, albiet, in not the most efficient way.

Updated unit tests to account for this new functionality.